rename: 컨벤션에 맞는 파일 이름 변경
