+++ src/components/common/Board/KBOBanner.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const KboBanner = () => {
   return (
-    <div className="w-full bg-[#002561] flex justify-center items-center py-16 px-16">
+    <div className="w-full bg-kbo-blue flex justify-center items-center py-16 px-16">
       <div className="w-[1100px] flex flex-col gap-6 text-white">
         <h2 className="text-4xl font-bold">KBO 통합 팬 커뮤니티</h2>
         <p className="text-base">함께 응원하고 소통하는 팬들의 공간입니다</p>
